kymux-types: Make ProtocolEndpoint dyn-compatible
The function ready(self) consumes self (as expected), but this makes the
trait dyn-non-compatible. As a consequence, it is not possible to call
ready() on a Box<dyn ProtocolEndpoint>, the concrete type is required.

To allow ProtocolEndpoint instances to be used dynamically, change the
function to accept a Box<Self> instead, and re-implement ready(self)
based on the boxed version for convenience.

This change will enable kycom to be decoupled from kyproto entirely.